% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_methods.R
\name{as.array,LatentNeuroVec-method}
\alias{as.array,LatentNeuroVec-method}
\title{Reconstruct LatentNeuroVec as a 4D array}
\usage{
\S4method{as.array}{LatentNeuroVec}(x, ...)
}
\arguments{
\item{x}{A \code{LatentNeuroVec} object.}

\item{...}{Ignored.}
}
\value{
A 4D numeric array with dimensions (x, y, z, time).
}
\description{
Reconstruct LatentNeuroVec as a 4D array
}
