% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_structure.R
\name{as.matrix.GroupDeltaLoadings}
\alias{as.matrix.GroupDeltaLoadings}
\title{Coerce group-plus-delta loadings to a dense matrix}
\usage{
\method{as.matrix}{GroupDeltaLoadings}(x, ...)
}
\arguments{
\item{x}{A `GroupDeltaLoadings` object.}

\item{...}{Ignored.}
}
\value{
Dense matrix of materialized loadings.
}
\description{
Coerce group-plus-delta loadings to a dense matrix
}
