% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar_wavelet.R
\name{as.matrix.HaarLatent}
\alias{as.matrix.HaarLatent}
\title{Convert HaarLatent to matrix}
\usage{
\method{as.matrix}{HaarLatent}(x, ...)
}
\arguments{
\item{x}{HaarLatent object}

\item{...}{Additional arguments (unused)}
}
\value{
Matrix of reconstructed values
}
\description{
Convert HaarLatent to matrix
}
