% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbf.R
\name{as_hrbf_latent}
\alias{as_hrbf_latent}
\title{Attach HRBF metadata to an existing LatentNeuroVec}
\usage{
as_hrbf_latent(lvec, params, centres = NULL, sigmas = NULL)
}
\arguments{
\item{lvec}{A LatentNeuroVec object}

\item{params}{HRBF parameters list}

\item{centres}{Optional matrix of HRBF centres}

\item{sigmas}{Optional vector of HRBF sigmas}
}
\value{
The LatentNeuroVec with HRBF metadata attached
}
\description{
Attach HRBF metadata to an existing LatentNeuroVec
}
