% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit_latent.R
\name{as_implicit_latent}
\alias{as_implicit_latent}
\title{Coerce an object to an ImplicitLatent decoder}
\usage{
as_implicit_latent(x, ...)
}
\arguments{
\item{x}{Object to coerce.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class `ImplicitLatent`.
}
\description{
Coerce an object to an ImplicitLatent decoder
}
