% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/latent_surface_vector.R,
%   R/latent_bilateral_surface_vector.R, R/transport_latent.R,
%   R/latent_block_vector.R
\name{basis_asset}
\alias{basis_asset}
\alias{basis_asset,LatentNeuroSurfaceVector-method}
\alias{basis_asset,BilatLatentNeuroSurfaceVector-method}
\alias{basis_asset,ImplicitLatent-method}
\alias{basis_asset,LatentNeuroVec-method}
\alias{basis_asset,BlockLatentNeuroVector-method}
\title{Extract the basis asset behind a latent object}
\usage{
basis_asset(x, ...)

\S4method{basis_asset}{LatentNeuroSurfaceVector}(x, ...)

\S4method{basis_asset}{BilatLatentNeuroSurfaceVector}(x, ...)

\S4method{basis_asset}{ImplicitLatent}(x, ...)

\S4method{basis_asset}{LatentNeuroVec}(x, ...)

\S4method{basis_asset}{BlockLatentNeuroVector}(x, ...)
}
\arguments{
\item{x}{A latent object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Basis asset object or \code{NULL}.
}
\description{
Extract the basis asset behind a latent object
}
