% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/awpt.R,
%   R/hierarchical_template.R, R/parcel_basis.R, R/surface_template.R
\name{basis_decoder}
\alias{basis_decoder}
\alias{basis_decoder,SurfaceAWPTBasisTemplate-method}
\alias{basis_decoder,AWPTBasisTemplate-method}
\alias{basis_decoder,HierarchicalBasisTemplate-method}
\alias{basis_decoder,ParcelBasisTemplate-method}
\alias{basis_decoder,SurfaceBasisTemplate-method}
\title{Extract a basis decoder from a template asset}
\usage{
basis_decoder(template, ...)

\S4method{basis_decoder}{SurfaceAWPTBasisTemplate}(template, ...)

\S4method{basis_decoder}{AWPTBasisTemplate}(template, ...)

\S4method{basis_decoder}{HierarchicalBasisTemplate}(template, ...)

\S4method{basis_decoder}{ParcelBasisTemplate}(template, ...)

\S4method{basis_decoder}{SurfaceBasisTemplate}(template, ...)
}
\arguments{
\item{template}{A template object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Basis decoder view object.
}
\description{
Extract a basis decoder from a template asset
}
