% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_wavelet.R
\name{basis_heat_wavelet}
\alias{basis_heat_wavelet}
\title{Heat wavelet basis specification}
\usage{
basis_heat_wavelet(scales = c(1, 2, 4, 8), order = 30, threshold = 1e-06)
}
\arguments{
\item{scales}{Numeric vector of heat diffusion scales.}

\item{order}{Polynomial approximation order.}

\item{threshold}{Threshold for small coefficients.}
}
\value{
A `spec_heat_wavelet` object for `lift()`.
}
\description{
Heat wavelet basis specification
}
