% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{BilatLatentNeuroSurfaceVector-class}
\alias{BilatLatentNeuroSurfaceVector-class}
\title{BilatLatentNeuroSurfaceVector Class}
\description{
An explicit bilateral surface latent representation composed of left and
right \code{LatentNeuroSurfaceVector} objects.
}
\section{Slots}{

\describe{
\item{\code{left}}{Left-hemisphere latent surface object.}

\item{\code{right}}{Right-hemisphere latent surface object.}

\item{\code{label}}{Character label.}

\item{\code{meta}}{Lightweight metadata list.}
}}

