% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_bilateral_surface_vector.R
\name{BilatLatentNeuroSurfaceVector}
\alias{BilatLatentNeuroSurfaceVector}
\title{Construct a bilateral surface latent object}
\usage{
BilatLatentNeuroSurfaceVector(left, right, label = "", meta = list())
}
\arguments{
\item{left}{Left \code{LatentNeuroSurfaceVector}.}

\item{right}{Right \code{LatentNeuroSurfaceVector}.}

\item{label}{Optional label.}

\item{meta}{Optional metadata list.}
}
\value{
A \code{BilatLatentNeuroSurfaceVector}.
}
\description{
Construct a bilateral surface latent object
}
