% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_block_vector.R
\name{BlockLatentNeuroVector}
\alias{BlockLatentNeuroVector}
\title{Construct a block-domain latent object}
\usage{
BlockLatentNeuroVector(blocks, label = "", meta = list())
}
\arguments{
\item{blocks}{Named list of explicit latent objects sharing a common basis.}

\item{label}{Optional label.}

\item{meta}{Optional metadata list.}
}
\value{
A \code{BlockLatentNeuroVector}.
}
\description{
Construct a block-domain latent object
}
