% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\docType{class}
\name{CoarsenedReduction-class}
\alias{CoarsenedReduction-class}
\title{Coarsened graph reduction (e.g., prolongation from coarse to fine)}
\description{
Coarsened graph reduction (e.g., prolongation from coarse to fine)
}
\section{Slots}{

\describe{
\item{\code{P_matrix}}{Sparse prolongation matrix (fine x coarse).}

\item{\code{coarse_adj}}{Optional sparse adjacency on the coarse graph.}
}}

