% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_methods.R
\name{concat,LatentNeuroVec,LatentNeuroVec-method}
\alias{concat,LatentNeuroVec,LatentNeuroVec-method}
\title{Concatenate LatentNeuroVec Objects}
\usage{
\S4method{concat}{LatentNeuroVec,LatentNeuroVec}(x, y, ...)
}
\arguments{
\item{x}{First \code{LatentNeuroVec}.}

\item{y}{Second \code{LatentNeuroVec}.}

\item{...}{Additional \code{LatentNeuroVec} objects to concatenate.}
}
\value{
A new \code{LatentNeuroVec} if all objects are compatible, otherwise
  a \code{\link[neuroim2]{NeuroVecSeq-class}}.
}
\description{
Concatenates two or more \code{LatentNeuroVec} objects along the temporal dimension.
}
\examples{
mask <- neuroim2::LogicalNeuroVol(
  array(TRUE, dim = c(2, 2, 1)),
  neuroim2::NeuroSpace(c(2, 2, 1))
)
lvec <- LatentNeuroVec(
  basis = matrix(1:6, nrow = 3),
  loadings = matrix(seq_len(8) / 10, nrow = 4),
  space = neuroim2::NeuroSpace(c(2, 2, 1, 3)),
  mask = mask,
  expect_dense = TRUE
)
combined <- concat(lvec, lvec)
dim(combined)

}
