% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar_wavelet.R
\name{get_morton_ordered_indices}
\alias{get_morton_ordered_indices}
\title{Morton-ordered voxel indices from a 3D mask}
\usage{
get_morton_ordered_indices(mask_3d_array, z_order_seed = 42L)
}
\description{
Morton-ordered voxel indices from a 3D mask
}
\keyword{internal}
