% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar_wavelet.R
\name{get_valid_finest_blocks}
\alias{get_valid_finest_blocks}
\title{Morton codes of valid finest-level blocks}
\usage{
get_valid_finest_blocks(mask_3d_array)
}
\description{
Morton codes of valid finest-level blocks
}
\keyword{internal}
