% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar_wavelet.R
\name{haar_meta}
\alias{haar_meta}
\title{Get metadata from Haar latent object}
\usage{
haar_meta(x)
}
\arguments{
\item{x}{A Haar latent object}
}
\value{
Metadata list or NULL
}
\description{
Get metadata from Haar latent object
}
