% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbf.R
\name{hrbf_meta}
\alias{hrbf_meta}
\title{Retrieve HRBF metadata if present}
\usage{
hrbf_meta(x)
}
\arguments{
\item{x}{A LatentNeuroVec object}
}
\value{
HRBF metadata list or NULL if not HRBF-tagged
}
\description{
Retrieve HRBF metadata if present
}
