% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit_latent.R
\name{implicit_latent}
\alias{implicit_latent}
\title{Construct an ImplicitLatent object}
\usage{
implicit_latent(coeff, decoder, meta, mask)
}
\arguments{
\item{coeff}{Arbitrary coefficient payload (list or matrix) needed by decoder.}

\item{decoder}{Function(time_idx = NULL, roi_mask = NULL, levels_keep = NULL) returning matrix.}

\item{meta}{List metadata; must include `family` string.}

\item{mask}{Logical 3D array (or LogicalNeuroVol) describing voxel support.}
}
\value{
An object of class `ImplicitLatent`.
}
\description{
Construct an ImplicitLatent object
}
