% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit_latent.R
\name{implicit_meta}
\alias{implicit_meta}
\title{Get metadata from ImplicitLatent object}
\usage{
implicit_meta(x)
}
\arguments{
\item{x}{An ImplicitLatent object}
}
\value{
Metadata list or NULL
}
\description{
Get metadata from ImplicitLatent object
}
