% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_template.R
\name{is_hierarchical_template}
\alias{is_hierarchical_template}
\title{Check whether an object is a HierarchicalBasisTemplate}
\usage{
is_hierarchical_template(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical scalar.
}
\description{
Check whether an object is a HierarchicalBasisTemplate
}
