% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbf.R
\name{is_hrbf_latent}
\alias{is_hrbf_latent}
\title{Check if latent object carries HRBF metadata}
\usage{
is_hrbf_latent(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if x has HRBF metadata
}
\description{
Check if latent object carries HRBF metadata
}
