% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit_latent.R
\name{is_implicit_latent}
\alias{is_implicit_latent}
\title{Test if object is an ImplicitLatent}
\usage{
is_implicit_latent(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if x is an ImplicitLatent
}
\description{
Test if object is an ImplicitLatent
}
