% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_structure.R
\name{is_shared_reference}
\alias{is_shared_reference}
\title{Test whether an object is a shared reference}
\usage{
is_shared_reference(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a shared reference
}
