% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_template.R
\name{is_surface_template}
\alias{is_surface_template}
\title{Test whether an object is a surface basis template}
\usage{
is_surface_template(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a surface basis template
}
