% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_template.R
\name{is_template}
\alias{is_template}
\title{Test whether an object is a supported template}
\usage{
is_template(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a supported template
}
