% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_latent.R
\name{is_transport_latent}
\alias{is_transport_latent}
\title{Test whether an object is a transport-backed latent object}
\usage{
is_transport_latent(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a transport-backed latent object
}
