% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/implicit_latent.R,
%   R/latent_surface_vector.R, R/latent_bilateral_surface_vector.R,
%   R/latent_block_vector.R, R/latent_methods.R
\name{latent_domain}
\alias{latent_domain}
\alias{latent_domain,ImplicitLatent-method}
\alias{latent_domain,LatentNeuroSurfaceVector-method}
\alias{latent_domain,BilatLatentNeuroSurfaceVector-method}
\alias{latent_domain,BlockLatentNeuroVector-method}
\alias{latent_domain,LatentNeuroVec-method}
\title{Extract the decoded domain associated with a latent object}
\usage{
latent_domain(x, ...)

\S4method{latent_domain}{ImplicitLatent}(x, ...)

\S4method{latent_domain}{LatentNeuroSurfaceVector}(x, ...)

\S4method{latent_domain}{BilatLatentNeuroSurfaceVector}(x, ...)

\S4method{latent_domain}{BlockLatentNeuroVector}(x, ...)

\S4method{latent_domain}{LatentNeuroVec}(x, ...)
}
\arguments{
\item{x}{A latent object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Domain object or identifier.
}
\description{
Extract the decoded domain associated with a latent object
}
