% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/implicit_latent.R,
%   R/latent_surface_vector.R, R/latent_bilateral_surface_vector.R,
%   R/latent_block_vector.R, R/latent_methods.R
\name{latent_meta}
\alias{latent_meta}
\alias{latent_meta,ImplicitLatent-method}
\alias{latent_meta,LatentNeuroSurfaceVector-method}
\alias{latent_meta,BilatLatentNeuroSurfaceVector-method}
\alias{latent_meta,BlockLatentNeuroVector-method}
\alias{latent_meta,LatentNeuroVec-method}
\title{Get lightweight metadata from a latent object}
\usage{
latent_meta(x, ...)

\S4method{latent_meta}{ImplicitLatent}(x, ...)

\S4method{latent_meta}{LatentNeuroSurfaceVector}(x, ...)

\S4method{latent_meta}{BilatLatentNeuroSurfaceVector}(x, ...)

\S4method{latent_meta}{BlockLatentNeuroVector}(x, ...)

\S4method{latent_meta}{LatentNeuroVec}(x, ...)
}
\arguments{
\item{x}{A latent object.}

\item{...}{Additional arguments (unused).}
}
\value{
Metadata list.
}
\description{
Get lightweight metadata from a latent object
}
