% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/implicit_latent.R,
%   R/latent_surface_vector.R, R/latent_bilateral_surface_vector.R,
%   R/latent_block_vector.R, R/latent_methods.R
\name{latent_support}
\alias{latent_support}
\alias{latent_support,ImplicitLatent-method}
\alias{latent_support,LatentNeuroSurfaceVector-method}
\alias{latent_support,BilatLatentNeuroSurfaceVector-method}
\alias{latent_support,BlockLatentNeuroVector-method}
\alias{latent_support,LatentNeuroVec-method}
\title{Extract the decoded support associated with a latent object}
\usage{
latent_support(x, ...)

\S4method{latent_support}{ImplicitLatent}(x, ...)

\S4method{latent_support}{LatentNeuroSurfaceVector}(x, ...)

\S4method{latent_support}{BilatLatentNeuroSurfaceVector}(x, ...)

\S4method{latent_support}{BlockLatentNeuroVector}(x, ...)

\S4method{latent_support}{LatentNeuroVec}(x, ...)
}
\arguments{
\item{x}{A latent object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Support object for the latent representation.
}
\description{
Extract the decoded support associated with a latent object
}
