% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_template.R
\name{load_hierarchical_template}
\alias{load_hierarchical_template}
\title{Load a hierarchical template from disk}
\usage{
load_hierarchical_template(file)
}
\arguments{
\item{file}{Path to .rds produced by save_hierarchical_template}
}
\value{
HierarchicalBasisTemplate
}
\description{
Load a hierarchical template from disk
}
