% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_template.R
\name{load_template}
\alias{load_template}
\title{Load a saved template from disk}
\usage{
load_template(file)
}
\arguments{
\item{file}{Path to a template RDS file.}
}
\value{
A supported template object.
}
\description{
Load a saved template from disk
}
