% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/latent_surface_vector.R,
%   R/latent_bilateral_surface_vector.R, R/latent_block_vector.R,
%   R/latent_neurovector.R
\name{loadings}
\alias{loadings}
\alias{loadings,LatentNeuroSurfaceVector-method}
\alias{loadings,BilatLatentNeuroSurfaceVector-method}
\alias{loadings,BlockLatentNeuroVector-method}
\alias{loadings,LatentNeuroVec-method}
\title{Get the loadings matrix (spatial components)}
\usage{
loadings(x, ...)

\S4method{loadings}{LatentNeuroSurfaceVector}(x)

\S4method{loadings}{BilatLatentNeuroSurfaceVector}(x)

\S4method{loadings}{BlockLatentNeuroVector}(x)

\S4method{loadings}{LatentNeuroVec}(x)
}
\arguments{
\item{x}{An object containing loadings (e.g., \code{LatentNeuroVec})}

\item{...}{Additional arguments (currently unused)}
}
\value{
The loadings matrix (typically voxels x components)
}
\description{
Extract the loadings matrix from a latent space representation. For
\code{LatentNeuroVec} objects, this returns the spatial loadings matrix
with dimensions (nVoxels x k) where k is the number of components and
nVoxels is the number of voxels within the mask.
}
\examples{
mask <- neuroim2::LogicalNeuroVol(
  array(TRUE, dim = c(2, 2, 1)),
  neuroim2::NeuroSpace(c(2, 2, 1))
)
lvec <- LatentNeuroVec(
  basis = matrix(1:6, nrow = 3),
  loadings = matrix(seq_len(8) / 10, nrow = 4),
  space = neuroim2::NeuroSpace(c(2, 2, 1, 3)),
  mask = mask,
  expect_dense = TRUE
)
l_matrix <- loadings(lvec)
dim(l_matrix)

}
