% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/latent_neurovector.R
\name{mask}
\alias{mask}
\alias{mask,LatentNeuroVec-method}
\title{Get the mask}
\usage{
\S4method{mask}{LatentNeuroVec}(x)
}
\arguments{
\item{x}{An object containing a mask (e.g., \code{LatentNeuroVec})}
}
\value{
The logical mask volume
}
\description{
Extract the brain mask from a latent space representation. For
\code{LatentNeuroVec} objects, this returns the \code{LogicalNeuroVol}
defining which voxels are included in the representation.
}
