% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_structure.R
\name{materialize_group_delta_loadings}
\alias{materialize_group_delta_loadings}
\title{Materialize group-plus-delta loadings}
\usage{
materialize_group_delta_loadings(x)
}
\arguments{
\item{x}{A `GroupDeltaLoadings` object.}
}
\value{
Matrix-like materialized loadings.
}
\description{
Materialize group-plus-delta loadings
}
