% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar_wavelet.R
\name{perform_haar_lift_analysis}
\alias{perform_haar_lift_analysis}
\title{Perform forward Haar lifting analysis}
\usage{
perform_haar_lift_analysis(
  data_matrix_T_x_Nmask,
  mask_3d_array,
  levels,
  z_order_seed = 42L
)
}
\description{
Perform forward Haar lifting analysis
}
\keyword{internal}
