% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar_wavelet.R
\name{perform_haar_lift_synthesis}
\alias{perform_haar_lift_synthesis}
\title{Perform inverse Haar lifting synthesis}
\usage{
perform_haar_lift_synthesis(
  coeff_list,
  mask_3d_array,
  levels,
  z_order_seed = 42L
)
}
\description{
Perform inverse Haar lifting synthesis
}
\keyword{internal}
