% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_roundtrip.R
\name{plot_benchmark_roundtrip}
\alias{plot_benchmark_roundtrip}
\title{Plot benchmark results}
\usage{
plot_benchmark_roundtrip(df)
}
\arguments{
\item{df}{Data frame from benchmark_roundtrip.}
}
\value{
ggplot object if ggplot2 available; otherwise prints df.
}
\description{
Plot benchmark results
}
