% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar_wavelet.R
\name{precompute_haar_scalings}
\alias{precompute_haar_scalings}
\title{Precompute scaling factors for Haar lifting}
\usage{
precompute_haar_scalings(mask_3d_array, levels)
}
\description{
Precompute scaling factors for Haar lifting
}
\keyword{internal}
