% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar_wavelet.R
\name{predict.HaarLatent}
\alias{predict.HaarLatent}
\title{Predict method for HaarLatent}
\usage{
\method{predict}{HaarLatent}(object, roi_mask = NULL, time_idx = NULL, levels_keep = NULL, ...)
}
\arguments{
\item{object}{HaarLatent object}

\item{roi_mask}{Optional ROI mask}

\item{time_idx}{Optional time indices}

\item{levels_keep}{Optional levels to keep}

\item{...}{Additional arguments passed to decoder}
}
\value{
Matrix of predicted values
}
\description{
Predict method for HaarLatent
}
