% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit_latent.R
\name{predict.ImplicitLatent}
\alias{predict.ImplicitLatent}
\title{Predict method for ImplicitLatent}
\usage{
\method{predict}{ImplicitLatent}(object, roi_mask = NULL, time_idx = NULL, levels_keep = NULL, ...)
}
\arguments{
\item{object}{ImplicitLatent object}

\item{roi_mask}{Optional ROI mask}

\item{time_idx}{Optional time indices}

\item{levels_keep}{Optional levels to keep}

\item{...}{Additional arguments (unused)}
}
\value{
Matrix of predicted values
}
\description{
Predict method for ImplicitLatent
}
