% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_basis.R
\name{print.ParcelBasisTemplate}
\alias{print.ParcelBasisTemplate}
\title{Print method for ParcelBasisTemplate}
\usage{
\method{print}{ParcelBasisTemplate}(x, ...)
}
\arguments{
\item{x}{A ParcelBasisTemplate object.}

\item{...}{Ignored.}
}
\value{
The input `x`, invisibly.
}
\description{
Print method for ParcelBasisTemplate
}
