% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_template.R
\name{project_hierarchical}
\alias{project_hierarchical}
\title{Project coefficients only (no LatentNeuroVec wrapper)}
\usage{
project_hierarchical(template, X)
}
\arguments{
\item{template}{HierarchicalBasisTemplate to use for projection}

\item{X}{matrix time x voxels (mask order) matching template mask}
}
\value{
Matrix of coefficients (time x atoms)
}
\description{
Project coefficients only (no LatentNeuroVec wrapper)
}
