% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_structure.R
\name{shared_reference_clear}
\alias{shared_reference_clear}
\title{Reset the in-session shared reference cache}
\usage{
shared_reference_clear()
}
\value{
Invisibly, `TRUE`.
}
\description{
Reset the in-session shared reference cache
}
