% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_spec.R
\name{spec_time_dct}
\alias{spec_time_dct}
\title{Temporal DCT spec}
\usage{
spec_time_dct(k = NULL, norm = c("ortho", "none"))
}
\arguments{
\item{k}{Optional number of components. If `NULL`, the encoder uses all
available DCT columns (`k = n_time`) at encode time.}

\item{norm}{Normalization ("ortho" or "none").}
}
\value{
A `spec_time_dct` object.
}
\description{
Temporal DCT spec
}
