% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/awpt.R,
%   R/hierarchical_template.R, R/parcel_basis.R, R/surface_template.R
\name{template_domain}
\alias{template_domain}
\alias{template_domain,SurfaceAWPTBasisTemplate-method}
\alias{template_domain,AWPTBasisTemplate-method}
\alias{template_domain,HierarchicalBasisTemplate-method}
\alias{template_domain,ParcelBasisTemplate-method}
\alias{template_domain,SurfaceBasisTemplate-method}
\title{Extract the domain associated with a template}
\usage{
template_domain(template, ...)

\S4method{template_domain}{SurfaceAWPTBasisTemplate}(template, ...)

\S4method{template_domain}{AWPTBasisTemplate}(template, ...)

\S4method{template_domain}{HierarchicalBasisTemplate}(template, ...)

\S4method{template_domain}{ParcelBasisTemplate}(template, ...)

\S4method{template_domain}{SurfaceBasisTemplate}(template, ...)
}
\arguments{
\item{template}{A template object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Domain object or identifier.
}
\description{
Extract the domain associated with a template
}
