% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/awpt.R,
%   R/hierarchical_template.R, R/parcel_basis.R, R/surface_template.R
\name{template_loadings}
\alias{template_loadings}
\alias{template_loadings,SurfaceAWPTBasisTemplate-method}
\alias{template_loadings,AWPTBasisTemplate-method}
\alias{template_loadings,HierarchicalBasisTemplate-method}
\alias{template_loadings,ParcelBasisTemplate-method}
\alias{template_loadings,SurfaceBasisTemplate-method}
\title{Extract template loadings}
\usage{
template_loadings(x, ...)

\S4method{template_loadings}{SurfaceAWPTBasisTemplate}(x, ...)

\S4method{template_loadings}{AWPTBasisTemplate}(x, ...)

\S4method{template_loadings}{HierarchicalBasisTemplate}(x, ...)

\S4method{template_loadings}{ParcelBasisTemplate}(x, ...)

\S4method{template_loadings}{SurfaceBasisTemplate}(x, ...)
}
\arguments{
\item{x}{A template object.}

\item{...}{Additional arguments (unused).}
}
\value{
Spatial loadings matrix stored by the template.
}
\description{
Extract template loadings
}
