% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hierarchical_template.R,
%   R/parcel_basis.R
\name{template_mask}
\alias{template_mask}
\alias{template_mask,HierarchicalBasisTemplate-method}
\alias{template_mask,ParcelBasisTemplate-method}
\title{Extract template mask}
\usage{
template_mask(x, ...)

\S4method{template_mask}{HierarchicalBasisTemplate}(x, ...)

\S4method{template_mask}{ParcelBasisTemplate}(x, ...)
}
\arguments{
\item{x}{A template object.}

\item{...}{Additional arguments (unused).}
}
\value{
Mask associated with the template.
}
\description{
Extract template mask
}
