% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/awpt.R,
%   R/hierarchical_template.R, R/parcel_basis.R, R/surface_template.R
\name{template_measure}
\alias{template_measure}
\alias{template_measure,SurfaceAWPTBasisTemplate-method}
\alias{template_measure,AWPTBasisTemplate-method}
\alias{template_measure,HierarchicalBasisTemplate-method}
\alias{template_measure,ParcelBasisTemplate-method}
\alias{template_measure,SurfaceBasisTemplate-method}
\title{Extract optional measure or mass information for a template}
\usage{
template_measure(template, ...)

\S4method{template_measure}{SurfaceAWPTBasisTemplate}(template, ...)

\S4method{template_measure}{AWPTBasisTemplate}(template, ...)

\S4method{template_measure}{HierarchicalBasisTemplate}(template, ...)

\S4method{template_measure}{ParcelBasisTemplate}(template, ...)

\S4method{template_measure}{SurfaceBasisTemplate}(template, ...)
}
\arguments{
\item{template}{A template object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Optional measure or mass object, or \code{NULL}.
}
\description{
Extract optional measure or mass information for a template
}
