% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hierarchical_template.R,
%   R/parcel_basis.R
\name{template_meta}
\alias{template_meta}
\alias{template_meta,HierarchicalBasisTemplate-method}
\alias{template_meta,ParcelBasisTemplate-method}
\title{Extract template metadata}
\usage{
template_meta(x, ...)

\S4method{template_meta}{HierarchicalBasisTemplate}(x, ...)

\S4method{template_meta}{ParcelBasisTemplate}(x, ...)
}
\arguments{
\item{x}{A template object.}

\item{...}{Additional arguments (unused).}
}
\value{
Metadata list.
}
\description{
Extract template metadata
}
