% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hierarchical_template.R,
%   R/parcel_basis.R
\name{template_project}
\alias{template_project}
\alias{template_project,HierarchicalBasisTemplate-method}
\alias{template_project,ParcelBasisTemplate-method}
\title{Project data onto a template}
\usage{
template_project(x, data, ...)

\S4method{template_project}{HierarchicalBasisTemplate}(x, data, ...)

\S4method{template_project}{ParcelBasisTemplate}(x, data, ...)
}
\arguments{
\item{x}{A template object.}

\item{data}{Numeric matrix (time x voxels in mask order).}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list with \code{coefficients} (time x atoms/components) and
  \code{offset} (voxel means or \code{numeric(0)}).
}
\description{
Project data onto a template
}
