% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/awpt.R,
%   R/hierarchical_template.R, R/parcel_basis.R, R/surface_template.R
\name{template_rank}
\alias{template_rank}
\alias{template_rank,SurfaceAWPTBasisTemplate-method}
\alias{template_rank,AWPTBasisTemplate-method}
\alias{template_rank,HierarchicalBasisTemplate-method}
\alias{template_rank,ParcelBasisTemplate-method}
\alias{template_rank,SurfaceBasisTemplate-method}
\title{Query the rank of a template basis}
\usage{
template_rank(template, ...)

\S4method{template_rank}{SurfaceAWPTBasisTemplate}(template, ...)

\S4method{template_rank}{AWPTBasisTemplate}(template, ...)

\S4method{template_rank}{HierarchicalBasisTemplate}(template, ...)

\S4method{template_rank}{ParcelBasisTemplate}(template, ...)

\S4method{template_rank}{SurfaceBasisTemplate}(template, ...)
}
\arguments{
\item{template}{A template object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Integer scalar rank.
}
\description{
Query the rank of a template basis
}
