% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/awpt.R,
%   R/hierarchical_template.R, R/parcel_basis.R, R/surface_template.R
\name{template_support}
\alias{template_support}
\alias{template_support,SurfaceAWPTBasisTemplate-method}
\alias{template_support,AWPTBasisTemplate-method}
\alias{template_support,HierarchicalBasisTemplate-method}
\alias{template_support,ParcelBasisTemplate-method}
\alias{template_support,SurfaceBasisTemplate-method}
\title{Extract the support associated with a template}
\usage{
template_support(template, ...)

\S4method{template_support}{SurfaceAWPTBasisTemplate}(template, ...)

\S4method{template_support}{AWPTBasisTemplate}(template, ...)

\S4method{template_support}{HierarchicalBasisTemplate}(template, ...)

\S4method{template_support}{ParcelBasisTemplate}(template, ...)

\S4method{template_support}{SurfaceBasisTemplate}(template, ...)
}
\arguments{
\item{template}{A template object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Support object describing the decoded support of the template.
}
\description{
Extract the support associated with a template
}
