% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bridge.R
\name{voxel_subset_to_gsp}
\alias{voxel_subset_to_gsp}
\title{Convert voxel subset to an rgsp graph}
\usage{
voxel_subset_to_gsp(mask, voxel_indices, k_neighbors = 6L)
}
\arguments{
\item{mask}{LogicalNeuroVol or 3D logical array.}

\item{voxel_indices}{Integer vector of voxel indices (mask order).}

\item{k_neighbors}{Number of nearest neighbours for k-NN graph.}
}
\value{
`gsp_graph` object (from rgsp).
}
\description{
Convert voxel subset to an rgsp graph
}
