% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/implicit_latent.R,
%   R/latent_surface_vector.R, R/latent_bilateral_surface_vector.R,
%   R/latent_block_vector.R, R/latent_methods.R
\name{wrap_decoded}
\alias{wrap_decoded}
\alias{wrap_decoded,ImplicitLatent-method}
\alias{wrap_decoded,LatentNeuroSurfaceVector-method}
\alias{wrap_decoded,BilatLatentNeuroSurfaceVector-method}
\alias{wrap_decoded,BlockLatentNeuroVector-method}
\alias{wrap_decoded,LatentNeuroVec-method}
\title{Wrap flat decoded outputs into a domain-native representation}
\usage{
wrap_decoded(x, values, time_idx = NULL, space = c("native", "template"), ...)

\S4method{wrap_decoded}{ImplicitLatent}(x, values, time_idx = NULL, space = c("native", "template"), ...)

\S4method{wrap_decoded}{LatentNeuroSurfaceVector}(x, values, time_idx = NULL, space = c("native", "template"), ...)

\S4method{wrap_decoded}{BilatLatentNeuroSurfaceVector}(x, values, time_idx = NULL, space = c("native", "template"), ...)

\S4method{wrap_decoded}{BlockLatentNeuroVector}(x, values, time_idx = NULL, space = c("native", "template"), ...)

\S4method{wrap_decoded}{LatentNeuroVec}(x, values, time_idx = NULL, space = c("native", "template"), ...)
}
\arguments{
\item{x}{A latent object.}

\item{values}{Flat decoded values, typically as a vector or matrix.}

\item{time_idx}{Optional integer time indices associated with \code{values}.}

\item{space}{Output space to wrap into.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Domain-native wrapped representation.
}
\description{
Wrap flat decoded outputs into a domain-native representation
}
