% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_methods.R
\name{as.matrix,LatentNeuroVec-method}
\alias{as.matrix,LatentNeuroVec-method}
\title{Reconstruct LatentNeuroVec as a matrix (time x voxels)}
\usage{
\S4method{as.matrix}{LatentNeuroVec}(x, ...)
}
\arguments{
\item{x}{A \code{LatentNeuroVec} object.}

\item{...}{Ignored.}
}
\value{
A numeric matrix with rows = time points, columns = mask voxels.
}
\description{
Reconstruct LatentNeuroVec as a matrix (time x voxels)
}
