% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_surface_vector.R,
%   R/latent_bilateral_surface_vector.R, R/latent_block_vector.R,
%   R/latent_methods.R
\name{as.matrix,LatentNeuroSurfaceVector-method}
\alias{as.matrix,LatentNeuroSurfaceVector-method}
\alias{as.matrix,BilatLatentNeuroSurfaceVector-method}
\alias{as.matrix,BlockLatentNeuroVector-method}
\alias{as.matrix,LatentNeuroVec-method}
\title{Reconstruct LatentNeuroVec as a matrix (time x voxels)}
\usage{
\S4method{as.matrix}{LatentNeuroSurfaceVector}(x, ...)

\S4method{as.matrix}{BilatLatentNeuroSurfaceVector}(x, ...)

\S4method{as.matrix}{BlockLatentNeuroVector}(x, ...)

\S4method{as.matrix}{LatentNeuroVec}(x, ...)
}
\arguments{
\item{x}{A \code{LatentNeuroVec} object.}

\item{...}{Ignored.}
}
\value{
A numeric matrix with rows = time points, columns = mask voxels.
}
\description{
Reconstruct LatentNeuroVec as a matrix (time x voxels)
}
