% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_haar_aliases.R
\name{as_haar_latent}
\alias{as_haar_latent}
\title{Convert to HaarLatent class}
\usage{
as_haar_latent(obj)
}
\arguments{
\item{obj}{Object to convert}
}
\value{
Object with HaarLatent class added
}
\description{
Convert to HaarLatent class
}
