% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/latent_surface_vector.R,
%   R/latent_bilateral_surface_vector.R, R/latent_block_vector.R,
%   R/latent_neurovector.R
\name{basis}
\alias{basis}
\alias{basis,LatentNeuroSurfaceVector-method}
\alias{basis,BilatLatentNeuroSurfaceVector-method}
\alias{basis,BlockLatentNeuroVector-method}
\alias{basis,LatentNeuroVec-method}
\title{Get the basis matrix (temporal components)}
\usage{
basis(x, ...)

\S4method{basis}{LatentNeuroSurfaceVector}(x)

\S4method{basis}{BilatLatentNeuroSurfaceVector}(x)

\S4method{basis}{BlockLatentNeuroVector}(x)

\S4method{basis}{LatentNeuroVec}(x)
}
\arguments{
\item{x}{An object containing a basis matrix (e.g., \code{LatentNeuroVec})}

\item{...}{Additional arguments (currently unused)}
}
\value{
The basis matrix (typically time x components)
}
\description{
Extract the basis matrix from a latent space representation. For
\code{LatentNeuroVec} objects, this returns the temporal basis matrix
with dimensions (nTime x k) where k is the number of components.
}
\examples{
\dontrun{
# For LatentNeuroVec:
b_matrix <- basis(lvec)
print(dim(b_matrix)) # nTime x nComponents
}

}
