% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{basis_slepian}
\alias{basis_slepian}
\alias{basis_pca}
\alias{basis_flat}
\title{Basis specifications (lightweight descriptors)}
\usage{
basis_slepian(k = 3, type = "laplacian")

basis_pca(k = 3, whiten = FALSE)

basis_flat()
}
\arguments{
\item{k}{Number of components.}

\item{type}{Basis flavor (implementation-dependent).}

\item{whiten}{Whether to whiten PCA scores.}
}
\description{
Basis specifications (lightweight descriptors)
}
