% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_spec}
\alias{encode_spec}
\title{Dispatch encoding based on spec type}
\usage{
encode_spec(x, spec, ...)
}
\arguments{
\item{x}{Data matrix.}

\item{spec}{Spec object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Encoded representation.
}
\description{
Dispatch encoding based on spec type
}
