% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrilatent-package.R
\docType{package}
\name{fmrilatent-package}
\alias{fmrilatent}
\alias{fmrilatent-package}
\title{fmrilatent: Latent Space Representations of fMRI Data}
\description{
Provides S4 classes and methods for representing neuroimaging data in latent spaces. The LatentNeuroVec class offers a memory-efficient representation using matrix factorization (basis x loadings), enabling compact storage and efficient computation for dimensionality-reduced fMRI data such as PCA or ICA decompositions. Integrates with the neuroim2 package for standard neuroimaging data structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/fmrilatent/}
  \item \url{https://github.com/bbuchsbaum/fmrilatent}
  \item Report bugs at \url{https://github.com/bbuchsbaum/fmrilatent/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com} (\href{https://orcid.org/0000-0002-1108-4866}{ORCID})

}
\keyword{internal}
