% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/implicit_latent.R,
%   R/latent_methods.R
\name{is_explicit_latent}
\alias{is_explicit_latent}
\alias{is_explicit_latent,ImplicitLatent-method}
\alias{is_explicit_latent,LatentNeuroVec-method}
\title{Test whether a latent object is explicit}
\usage{
is_explicit_latent(x, ...)

\S4method{is_explicit_latent}{ImplicitLatent}(x, ...)

\S4method{is_explicit_latent}{LatentNeuroVec}(x, ...)
}
\arguments{
\item{x}{A latent object.}

\item{...}{Additional arguments (unused).}
}
\value{
Logical scalar; \code{TRUE} when the object stores explicit basis and
  loadings matrices.
}
\description{
Test whether a latent object is explicit
}
