% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_haar_aliases.R
\name{is_haar_latent}
\alias{is_haar_latent}
\title{Test if object is a Haar latent representation}
\usage{
is_haar_latent(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if x is a Haar latent
}
\description{
Test if object is a Haar latent representation
}
