% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/latent_neurovector.R
\name{loadings}
\alias{loadings}
\alias{loadings,LatentNeuroVec-method}
\title{Get the loadings matrix (spatial components)}
\usage{
loadings(x, ...)

\S4method{loadings}{LatentNeuroVec}(x)
}
\arguments{
\item{x}{An object containing loadings (e.g., \code{LatentNeuroVec})}

\item{...}{Additional arguments (currently unused)}
}
\value{
The loadings matrix (typically voxels x components)
}
\description{
Extract the loadings matrix from a latent space representation. For
\code{LatentNeuroVec} objects, this returns the spatial loadings matrix
with dimensions (nVoxels x k) where k is the number of components and
nVoxels is the number of voxels within the mask.
}
\examples{
\dontrun{
# For LatentNeuroVec:
l_matrix <- loadings(lvec)
print(dim(l_matrix)) # nVoxels x nComponents
}

}
