% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/latent_neurovector.R
\name{offset}
\alias{offset}
\alias{offset,LatentNeuroVec-method}
\title{Get the offset vector}
\usage{
\S4method{offset}{LatentNeuroVec}(object)
}
\arguments{
\item{object}{An object containing an offset (e.g., \code{LatentNeuroVec})}
}
\value{
The offset vector (length = number of voxels in mask)
}
\description{
Extract the offset vector from a latent space representation. For
\code{LatentNeuroVec} objects, this returns the voxel-wise offset
(mean or intercept) that is added after the basis x loadings reconstruction.
}
\examples{
\dontrun{
# For LatentNeuroVec:
off_vector <- offset(lvec)
print(length(off_vector)) # Should equal number of voxels in mask
}

}
