% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_template.R
\name{print.SurfaceBasisTemplate}
\alias{print.SurfaceBasisTemplate}
\title{Print method for SurfaceBasisTemplate}
\usage{
\method{print}{SurfaceBasisTemplate}(x, ...)
}
\arguments{
\item{x}{A \code{SurfaceBasisTemplate}.}

\item{...}{Ignored.}
}
\description{
Print method for SurfaceBasisTemplate
}
