% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_template.R
\name{save_hierarchical_template}
\alias{save_hierarchical_template}
\title{Save a hierarchical template to disk}
\usage{
save_hierarchical_template(template, file, compress = "xz")
}
\arguments{
\item{template}{HierarchicalBasisTemplate}

\item{file}{Path to .rds file}

\item{compress}{Compression passed to saveRDS (default "xz")}
}
\description{
Save a hierarchical template to disk
}
