% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hierarchical_template.R,
%   R/parcel_basis.R
\name{save_template}
\alias{save_template}
\alias{save_template,HierarchicalBasisTemplate-method}
\alias{save_template,ParcelBasisTemplate-method}
\title{Save a template object to disk}
\usage{
save_template(template, file, compress = "xz", ...)

\S4method{save_template}{HierarchicalBasisTemplate}(template, file, compress = "xz", ...)

\S4method{save_template}{ParcelBasisTemplate}(template, file, compress = "xz", ...)
}
\arguments{
\item{template}{A template object.}

\item{file}{Output path.}

\item{compress}{Compression passed to \code{saveRDS()}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Normalized output path, invisibly.
}
\description{
Save a template object to disk
}
