% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{spec_hierarchical_template}
\alias{spec_hierarchical_template}
\title{Create hierarchical template spec}
\usage{
spec_hierarchical_template(template = NULL, template_file = NULL)
}
\arguments{
\item{template}{HierarchicalBasisTemplate object}

\item{template_file}{Path to saved template file}
}
\value{
Spec object of class spec_hierarchical
}
\description{
Create hierarchical template spec
}
