% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_spec.R
\name{spec_space_heat}
\alias{spec_space_heat}
\title{Spatial heat-wavelet spec (graph diffusion)}
\usage{
spec_space_heat(
  scales = c(1, 2, 4, 8),
  order = 30L,
  threshold = 1e-06,
  k_neighbors = 6L
)
}
\arguments{
\item{scales}{Heat scales.}

\item{order}{Polynomial order.}

\item{threshold}{Threshold for small coefficients.}

\item{k_neighbors}{k-NN graph parameter.}
}
\value{
A `spec_space_heat` object.
}
\description{
Spatial heat-wavelet spec (graph diffusion)
}
