% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{spec_space_hrbf}
\alias{spec_space_hrbf}
\title{Spatial HRBF spec}
\usage{
spec_space_hrbf(params = list())
}
\arguments{
\item{params}{HRBF parameter list (sigma0, levels, radius_factor, kernel_type, seed).}
}
\value{
A `spec_space_hrbf` object.
}
\description{
Spatial HRBF spec
}
