% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{spec_space_slepian}
\alias{spec_space_slepian}
\title{Spatial Slepian spec}
\usage{
spec_space_slepian(k = 3L, k_neighbors = 6L)
}
\arguments{
\item{k}{Components per cluster.}

\item{k_neighbors}{k-NN graph parameter.}
}
\value{
A `spec_space_slepian` object.
}
\description{
Spatial Slepian spec
}
