% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{spec_space_wavelet_active}
\alias{spec_space_wavelet_active}
\title{Spatial wavelet (active pencil) spec}
\usage{
spec_space_wavelet_active(levels_space = 2L, levels_time = 0L, threshold = 0)
}
\arguments{
\item{levels_space}{Spatial lifting levels.}

\item{levels_time}{Optional time lifting levels.}

\item{threshold}{Threshold after space transform.}
}
\value{
A `spec_space_wavelet_active` object.
}
\description{
Spatial wavelet (active pencil) spec
}
