% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{spec_st}
\alias{spec_st}
\title{Spatiotemporal spec (separable)}
\usage{
spec_st(time, space, core_mode = c("auto", "explicit"))
}
\arguments{
\item{time}{Temporal spec (`spec_time_*`).}

\item{space}{Spatial spec (`spec_space_*`).}

\item{core_mode}{Reserved (currently "auto").}
}
\value{
A `spec_st` object.
}
\description{
Spatiotemporal spec (separable)
}
