% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{spec_time_bspline}
\alias{spec_time_bspline}
\title{Temporal B-spline spec}
\usage{
spec_time_bspline(
  k,
  degree = 3L,
  include_intercept = FALSE,
  orthonormalize = TRUE
)
}
\arguments{
\item{k}{Components (df).}

\item{degree}{Spline degree (default 3).}

\item{include_intercept}{Logical include intercept.}

\item{orthonormalize}{Logical orthonormalize columns (default TRUE).}
}
\value{
A `spec_time_bspline` object.
}
\description{
Temporal B-spline spec
}
